# api/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Geocars API",
        "description": "API documentation for the Geocars application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:\/\/localhost\/api\/v1",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "\/api\/CarRental": {
            "get": {
                "tags": [
                    "CarRental"
                ],
                "summary": "Get paginated list of CarRental",
                "description": "Retrieve a paginated list of CarRental with optional search",
                "operationId": "getCarRentalPaginated",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    },
                    {
                        "name": "rows",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/PaginatedCarRentalResponse200"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CarRental"
                ],
                "summary": "Create a new CarRental",
                "description": " Create a new CarRental with the provided details",
                "operationId": "createCarRental",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/CarRental"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CarRental created successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/CreateCarRentalResponse200"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/CarRental\/{id}": {
            "get": {
                "tags": [
                    "CarRental"
                ],
                "summary": "Get a specific CarRental",
                "description": "Retrieve a CarRental by its ID",
                "operationId": "getCarRentalById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/GetCarRentalResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CarRental not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "CarRental"
                ],
                "summary": "Update a CarRental",
                "description": "Update an existing CarRental with the provided details",
                "operationId": "updateCarRental",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "$ref": "#\/components\/schemas\/CarRental"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CarRental updated successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/UpdateCarRentalResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CarRental not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CarRental"
                ],
                "summary": "Delete a CarRental",
                "description": "Delete a CarRental by its ID",
                "operationId": "deleteCarRental",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "CarRental deleted successfully",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/DeleteCarRentalResponse200"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CarRental not found"
                    },
                    "500": {
                        "description": " Internal server error",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "$ref": "#\/components\/schemas\/InternalServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}